+++ packages/sampler/.storybook/addons.js
@@ -1,3 +1,1 @@
-import '@kadira/storybook/addons';
-import 'react-storybook-addon-backgrounds/register';
-import '@kadira/storybook-addon-knobs/register';
+import '../src/addons';
