+++ packages/sampler/src/addons.js
@@ -1,0 +1,3 @@
+import '@kadira/storybook/addons';
+import 'react-storybook-addon-backgrounds/register';
+import '@kadira/storybook-addon-knobs/register';
